Extracting `StreamSelect` component.
