+++ graylog2-web-interface/src/components/configurations/decorators/DecoratorsConfigUpdate.jsx
@@ -2,15 +2,10 @@
 import React, { useCallback, useState } from 'react';
 import { groupBy } from 'lodash';
 
-import Select from 'components/common/Select';
 import DecoratorList from 'views/components/messagelist/decorators/DecoratorList';
-import { defaultCompare } from 'views/logic/DefaultCompare';
 import Decorator from 'views/components/messagelist/decorators/Decorator';
 import AddDecoratorButton from 'views/components/messagelist/decorators/AddDecoratorButton';
-import SelectContainer from './SelectContainer';
-
-const DEFAULT_STREAM_ID = '000000000000000000000001';
-const DEFAULT_SEARCH_ID = 'DEFAULT_SEARCH';
+import StreamSelect, { DEFAULT_SEARCH_ID, DEFAULT_STREAM_ID } from './StreamSelect';
 
 const _formatDecorator = (decorator, decorators, decoratorTypes, updateFn) => {
   const typeDefinition = decoratorTypes[decorator.type] || { requested_configuration: {}, name: `Unknown type: ${decorator.type}` };
@@ -50,24 +45,12 @@ const DecoratorsConfigUpdate = ({ streams, decorators, types }: Props) => {
 
   const nextOrder = sortedDecorators.reduce((currentMax, decorator) => Math.max(currentMax, decorator.order), 0) + 1;
 
-  const options = [{ label: 'Default Search', value: DEFAULT_SEARCH_ID }, ...streams
-    .sort(({ title: key1 }, { title: key2 }) => defaultCompare(key1, key2))
-    .map(({ title, id }) => ({ label: title, value: id }))];
-
   const onCreate = useCallback(newDecorator => setUpdatedDecorators([...updatedDecorators, newDecorator]), [updatedDecorators, setUpdatedDecorators]);
 
   return (
     <React.Fragment>
       <p>Select the stream for which you want to change the set of default decorators.</p>
-      <SelectContainer>
-        <Select inputId="streams-filter"
-                onChange={setCurrentStream}
-                options={options}
-                clearable={false}
-                style={{ width: '100%' }}
-                placeholder="There are no decorators configured for any stream."
-                value={currentStream} />
-      </SelectContainer>
+      <StreamSelect onChange={setCurrentStream} value={currentStream} streams={streams} />
 
       <p>Select the type to create a new decorator for this stream:</p>
       <AddDecoratorButton stream={currentStream} nextOrder={nextOrder} decoratorTypes={types} onCreate={onCreate} showHelp={false} />
